Deduplicate artifact name on collisions. Store artifact relative path.
Summary:
Artifacts are expected to have unique names within a bucket. So, if two artifacts legitimately have
the same name, only the first one gets created on the store and the second one fails to get created.
To be more user friendly, we automatically deduplicate the artifact name (with a unique random suffix to
maintain the uniqueness constraint while leaving the artifact URL human-readable. Also, to avoid
losing original file name (and path), we now support a 'relativePath' field (which defaults to the
original artifact name if not explicitly set).

Test Plan: Added/modified unit and integration tests

Reviewers: kylec

Reviewed By: kylec

Subscribers: changesbot, jukka, nate

Differential Revision: https://tails.corp.dropbox.com/D139553