Don't query bucket while appending logchunk.
Summary:
If we're appending to a logchunk, we only need to verify that the corresponding artifact
exists. There is no value in checking if its bucket is present (if either bucket or artifact is not
found, return 404). This reduces one DB query for an API route which constitues 40-80% of all AS
traffic.

Test Plan: Integration tests continue to pass

Reviewers: kylec

Reviewed By: kylec

Subscribers: changesbot, nate

Differential Revision: https://tails.corp.dropbox.com/D160589